hello: rename contracts to universal and connected
