+++ examples/hello/contracts/Universal.sol
@@ -6,7 +6,7 @@ import "@zetachain/protocol-contracts/contracts/zevm/interfaces/UniversalContrac
 import "@zetachain/protocol-contracts/contracts/zevm/interfaces/IGatewayZEVM.sol";
 import "@zetachain/protocol-contracts/contracts/zevm/GatewayZEVM.sol";
 
-contract Hello is UniversalContract {
+contract Universal is UniversalContract {
     GatewayZEVM public immutable gateway;
 
     event HelloEvent(string, string);
