+++ examples/hello/tasks/universalWithdrawAndCall.ts
@@ -56,17 +56,34 @@ const main = async (args: any, hre: HardhatRuntimeEnvironment) => {
   );
 
   const gasLimit = hre.ethers.BigNumber.from(args.txOptionsGasLimit);
-  const zrc20 = new ethers.Contract(args.zrc20, ZRC20ABI.abi, signer);
-  const [, gasFee] = await zrc20.withdrawGasFeeWithGasLimit(gasLimit);
-  const zrc20TransferTx = await zrc20.approve(args.contract, gasFee, txOptions);
 
-  await zrc20TransferTx.wait();
+  const amount = hre.ethers.utils.parseUnits(args.amount, 18);
+
+  const zrc20 = new ethers.Contract(args.zrc20, ZRC20ABI.abi, signer);
+  const [gasZRC20, gasFee] = await zrc20.withdrawGasFeeWithGasLimit(gasLimit);
+  const gasZRC20Contract = new ethers.Contract(gasZRC20, ZRC20ABI.abi, signer);
+  const gasFeeApprove = await gasZRC20Contract.approve(
+    args.contract,
+    gasZRC20 == args.zrc20 ? gasFee.add(amount) : gasFee,
+    txOptions
+  );
+  await gasFeeApprove.wait();
+
+  if (gasZRC20 !== args.zrc20) {
+    const targetTokenApprove = await zrc20.approve(
+      args.contract,
+      gasFee.add(amount),
+      txOptions
+    );
+    await targetTokenApprove.wait();
+  }
 
   const factory = (await hre.ethers.getContractFactory(args.name)) as any;
   const contract = factory.attach(args.contract);
 
-  const tx = await contract.call(
+  const tx = await contract.withdrawAndCall(
     ethers.utils.hexlify(args.receiver),
+    amount,
     args.zrc20,
     message,
     gasLimit,
@@ -80,8 +97,8 @@ const main = async (args: any, hre: HardhatRuntimeEnvironment) => {
 };
 
 task(
-  "hello-call",
-  "Calls the gatewayCall function on a contract on ZetaChain",
+  "universal-withdraw-and-call",
+  "Withdraw ZRC-20 and call a function on a connected chain",
   main
 )
   .addParam("contract", "The address of the deployed Hello contract")
@@ -116,6 +133,7 @@ task(
     types.int
   )
   .addParam("function", `Function to call (example: "hello(string)")`)
-  .addParam("name", "The name of the contract", "Hello")
+  .addParam("name", "The name of the contract", "Universal")
+  .addParam("amount", "Amount of ZRC-20 to withdraw")
   .addParam("types", `The types of the parameters (example: '["string"]')`)
   .addVariadicPositionalParam("values", "The values of the parameters");
